Lots of updates for the Hera Communication Interface
