+++ data/src/main/java/hera/database/JPAUtil.java
@@ -10,42 +10,42 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class JPAUtil {
-  private static final Logger LOG = LoggerFactory.getLogger(JPAUtil.class);
+	private static final Logger LOG = LoggerFactory.getLogger(JPAUtil.class);
 
-  private static final String PERSISTENCE_UNIT_NAME = "HERA";
+	private static final String PERSISTENCE_UNIT_NAME = "HERA";
 
-  private static EntityManagerFactory factory;
+	private static EntityManagerFactory factory;
 
-  static EntityManager getEntityManager() {
-    if (factory == null) {
-      Map<String, String> env = System.getenv();
-      Map<String, Object> configOverrides = new HashMap<>();
-      for (String envName : env.keySet()) {
-        switch(envName) {
-          case "HERA_JDBC_USER":
-            configOverrides.put("javax.persistence.jdbc.user", env.get(envName));
-            break;
-          case "HERA_JDBC_PWD":
-            configOverrides.put("javax.persistence.jdbc.password", env.get(envName));
-            break;
-          case "HERA_JDBC_URL":
-            configOverrides.put("javax.persistence.jdbc.url", env.get(envName));
-            break;
-        }
-      }
+	static EntityManager getEntityManager() {
+		if (factory == null) {
+			Map<String, String> env = System.getenv();
+			Map<String, Object> configOverrides = new HashMap<>();
+			for (String envName : env.keySet()) {
+				switch(envName) {
+					case "HERA_DB_USER":
+						configOverrides.put("javax.persistence.jdbc.user", env.get(envName));
+						break;
+					case "HERA_DB_PWD":
+						configOverrides.put("javax.persistence.jdbc.password", env.get(envName));
+						break;
+					case "HERA_DB_URL":
+						configOverrides.put("javax.persistence.jdbc.url", env.get(envName));
+						break;
+				}
+			}
 
-      LOG.info("Creating DB connection");
-      factory = Persistence.createEntityManagerFactory(PERSISTENCE_UNIT_NAME, configOverrides);
-      LOG.info("EntityManagerFactory {} created", PERSISTENCE_UNIT_NAME);
-      LOG.info("Successfully connected to DB");
-    }
-    return factory.createEntityManager();
-  }
+			LOG.info("Creating DB connection");
+			factory = Persistence.createEntityManagerFactory(PERSISTENCE_UNIT_NAME, configOverrides);
+			LOG.info("EntityManagerFactory {} created", PERSISTENCE_UNIT_NAME);
+			LOG.info("Successfully connected to DB");
+		}
+		return factory.createEntityManager();
+	}
 
-  static void shutdownFactory() {
-    if (factory != null) {
-      factory.close();
-      LOG.info("EntityManagerFactory {} closed", PERSISTENCE_UNIT_NAME);
-    }
-  }
+	static void shutdownFactory() {
+		if (factory != null) {
+			factory.close();
+			LOG.info("EntityManagerFactory {} closed", PERSISTENCE_UNIT_NAME);
+		}
+	}
 }
